refactor: 커맨드 애플리케이션 로직을 각 유즈케이스에 맞게 클래스 분리
