+++ src/main/java/com/prography/lighton/auth/application/service/MemberWithdrawalService.java
@@ -4,7 +4,6 @@ import com.prography.lighton.artist.users.application.service.ArtistService;
 import com.prography.lighton.member.common.domain.entity.Member;
 import com.prography.lighton.member.common.infrastructure.repository.MemberRepository;
 import com.prography.lighton.member.common.infrastructure.repository.TemporaryMemberRepository;
-import com.prography.lighton.member.users.application.UserMemberCommandService;
 import com.prography.lighton.performance.users.application.service.UserPerformanceLikeService;
 import com.prography.lighton.performance.users.application.service.UserPerformanceService;
 import lombok.RequiredArgsConstructor;
