+++ src/test/java/com/prography/lighton/auth/application/service/MemberWithdrawalServiceTest.java
@@ -8,7 +8,6 @@ import com.prography.lighton.common.fixture.MemberTestFixture;
 import com.prography.lighton.member.common.domain.entity.Member;
 import com.prography.lighton.member.common.infrastructure.repository.MemberRepository;
 import com.prography.lighton.member.common.infrastructure.repository.TemporaryMemberRepository;
-import com.prography.lighton.member.users.application.UserMemberCommandService;
 import com.prography.lighton.performance.users.application.service.UserPerformanceLikeService;
 import com.prography.lighton.performance.users.application.service.UserPerformanceService;
 import org.junit.jupiter.api.DisplayName;
