+++ src/test/java/com/prography/lighton/common/presentation/ControllerTest.java
@@ -11,8 +11,7 @@ import com.prography.lighton.auth.presentation.AuthController;
 import com.prography.lighton.genre.infrastructure.cache.GenreCache;
 import com.prography.lighton.member.common.infrastructure.repository.MemberRepository;
 import com.prography.lighton.member.common.infrastructure.repository.PreferredGenreRepository;
-import com.prography.lighton.member.users.application.UserMemberCommandService;
-import com.prography.lighton.member.users.application.UserMemberQueryService;
+import com.prography.lighton.member.users.application.query.UserMemberQueryService;
 import com.prography.lighton.member.users.presentation.UserMemberController;
 import com.prography.lighton.region.infrastructure.cache.RegionCache;
 import org.springframework.beans.factory.annotation.Autowired;
